+++ driving_log_replayer_analyzer/driving_log_replayer_analyzer/data/__init__.py
@@ -1,0 +1,78 @@
+# Copyright (c) 2022 TierIV.inc
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+from dataclasses import dataclass
+import math
+from typing import Dict
+
+
+@dataclass
+class Stamp:
+    timestamp_system: float = -1.0
+    timestamp_ros: float = -1.0
+
+    def __init__(self, json_dict: Dict) -> None:
+        try:
+            self.timestamp_system = json_dict["Stamp"]["System"]
+            self.timestamp_ros = json_dict["Stamp"]["ROS"]
+        except (KeyError, IndexError):
+            pass
+
+    def validate(self):
+        return self.timestamp_system > 0.0 and self.timestamp_ros > 0.0
+
+
+@dataclass
+class Position:
+    # In the vehicle's coordinate system, x is frontal and y is lateral, but in the graph, x is lateral and y is vertical.
+    # Swap axes to match the xy in the graph.
+    x: float = None
+    y: float = None
+    z: float = None
+
+    def __init__(self, data: Dict = {}) -> None:
+        self.try_parse_dict(data)
+
+    def try_parse_dict(self, data):
+        if isinstance(data, dict):
+            try:
+                self.y = data["x"]  # Swap axes
+                self.x = data["y"]  # Swap axes
+                self.z = data["z"]
+            except KeyError:
+                pass
+        elif isinstance(data, list):
+            try:
+                self.y = data[0]  # Swap axes
+                self.x = data[1]  # Swap axes
+                self.z = data[2]
+            except IndexError:
+                pass
+        else:
+            raise NotImplementedError("Input data should be a dict or list.")
+
+    def validate(self):
+        return self.x is not None and self.y is not None and self.z is not None
+
+    def get_xy_distance(self) -> float:
+        if self.validate():
+            return math.hypot(self.x, self.y)
+        else:
+            return None
+
+    def add_overhang(self, val: float):
+        self.y = self.y + val
+
+    def sub_overhang(self, val: float):
+        self.y = self.y - val
